+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Allow pass options to `take()` via driver API.
 - Optional `consumer_group(opts, task)` driver API for routing waiting
   consumers to matching tasks.
-
+- Support extra statistics from driver API.
 
 ### Fixed